# add more refinement support for inference and literals