--- a/typed-racket-lib/typed-racket/typecheck/tc-funapp.rkt
+++ b/typed-racket-lib/typed-racket/typecheck/tc-funapp.rkt
@@ -111,13 +111,20 @@
                            [drest
                             (infer/dots
                              fixed-vars dotted-var argtys dom (car drest) rng (fv rng)
-                             #:expected (and expected (tc-results->values expected)))]
+                             #:expected (and expected (tc-results->values expected))
+                             #:objs argobjs)]
                            [rest
                             (infer/vararg fixed-vars (list dotted-var) argtys dom rest rng
-                                          (and expected (tc-results->values expected)))]
+                                          (and expected (tc-results->values expected))
+                                          #:objs argobjs)]
                            ;; no rest or drest
-                           [else (infer fixed-vars (list dotted-var) argtys dom rng
-                                        (and expected (tc-results->values expected)))])))
+                           [else (infer fixed-vars
+                                        (list dotted-var)
+                                        argtys
+                                        dom
+                                        rng
+                                        (and expected (tc-results->values expected))
+                                        #:objs argobjs)])))
          #:function-type f-type
          #:args-results args-res
          #:expected expected)]
@@ -141,7 +148,8 @@
          (let ([rng (subst-dom-objs argtys argobjs rng)])
            (extend-tvars vars
                          (infer/vararg vars null argtys dom rest rng
-                                       (and expected (tc-results->values expected)))))
+                                       (and expected (tc-results->values expected))
+                                       #:objs argobjs)))
          #:function-type f-type
          #:args-results args-res
          #:expected expected)]