--- a/typed-racket-lib/typed-racket/types/subtract.rkt
+++ b/typed-racket-lib/typed-racket/types/subtract.rkt
@@ -12,14 +12,14 @@
 ;; Type Type -> Type
 ;; conservatively calculates set subtraction
 ;; between the types (i.e. t - s)
-(define (subtract t s #:obj [obj -empty-obj])
+(define (subtract t s [obj -empty-obj])
   (define s-mask (mask s))
   (define result
     (let recurse ([t t] [obj obj])
       (define (sub t [obj -empty-obj]) (recurse t obj))
       (match t
         [_ #:when (disjoint-masks? (mask t) s-mask) t]
-        [_ #:when (subtype t s #:obj obj) -Bottom]
+        [_ #:when (subtype t s obj) -Bottom]
         [(or (App: _ _) (? Name?))
          ;; must be different, since they're not subtypes
          ;; and n must refer to a distinct struct type
@@ -43,5 +43,5 @@
         [(Poly: vs b) (make-Poly vs (sub b) obj)]
         [_ t])))
   (cond
-    [(subtype t result #:obj obj) t]
+    [(subtype t result obj) t]
     [else result]))