This PR adds more support for refinement reasoning, in particular type inference is now aware of argument objects which allows for more programs w/ refinements to typecheck. Additionally, working with vector types and literals that are refined or need to have properties about their length proven now works.

I need to add tests and peruse the changes more, but what is here at the moment seems to work and is representative of the scope I intend for this PR.

- [x] Add tests
- [x] more QC
- [x] sanity check that perf isn't affected (it shouldn't be w/ these changes)

Next stop (i.e. PR), function types that support inter-argument dependency!