@pnwamk 2017-04-04
Current performance results for this PR compaired to the Apr 4 snapshot =\
```
schml-specify-rep: 0.11% faster (8.72s to 8.71s)
schml-interp-casts-help: 0.85% faster (21.24s to 21.06s)
parser: 4.08% slower (1.96s to 2.04s)
old-metrics: 3.85% slower (2.34s to 2.43s)
new-metrics: 11.64% slower (2.92s to 3.26s)
math-flonum: 4.5% slower (3.78s to 3.95s)
fsm: 2.78% slower (7.56s to 7.77s)
forth: 3.51% slower (12.82s to 13.27s)
dungeon: 2.24% slower (14.28s to 14.6s)
acquire: 1.97% slower (36.54s to 37.26s)
```