@pnwamk 2017-04-04
This needs some further investigation w.r.t. perf