@pnwamk 2017-04-04
tinkered around w/ some things for performance sake, more or less same story:

```
schml-specify-rep: 0.81% faster (8.59s to 8.52s)
schml-interp-casts-help: 0.76% faster (21.01s to 20.85s)
parser: 4.17% slower (1.92s to 2.0s)
old-metrics: 2.6% slower (2.31s to 2.37s)
new-metrics: 6.9% slower (2.9s to 3.1s)
math-flonum: 6.25% slower (3.68s to 3.91s)
fsm: 1.2% slower (7.51s to 7.6s)
forth: 3.26% slower (12.88s to 13.3s)
dungeon: 2.16% slower (14.32s to 14.63s)
acquire: 0.82% slower (37.64s to 37.95s)
```