@pnwamk 2017-04-05
little more tinkering, less of a perf impact now:
```
schml-specify-rep: 2.62% faster (8.78s to 8.55s)
schml-interp-casts-help: 1.31% faster (21.45s to 21.17s)
parser: 2.55% slower (1.96s to 2.01s)
old-metrics: 1.27% slower (2.37s to 2.4s)
new-metrics: 4.35% slower (2.99s to 3.12s)
math-flonum: 3.47% slower (3.75s to 3.88s)
fsm: 1.19% slower (7.58s to 7.67s)
forth: 2.76% slower (13.03s to 13.39s)
dungeon: 1.78% slower (14.61s to 14.87s)
acquire: 0.26% slower (37.94s to 38.04s)
```