@samth 2017-04-05
Did the kw argument change really affect things?