@pnwamk 2017-04-05
Yes -- I just now reverted that file to the kw version and double checked how `new-metrics.rkt` is affected on my machine:

avg w/ keyword obj arg: 2.59s

avg w/ optional obj arg: 2.48s